agent_bridge/pyproject.toml
[project]
name = "agent-bridge"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
    "mcp>=1.0",
    "uvicorn>=0.30",
    "starlette>=0.38",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio", "anyio[trio]", "httpx>=0.27"]

[project.scripts]
agent-bridge = "agent_bridge.presentation.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_bridge"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
